Backend/requirements.txt
Django==4.0
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
gunicorn==20.0.4